option for 32/64 bits

+++ install-linux-dev-vm.sh
@@ -1,11 +1,19 @@
 #!/bin/bash
+clear
 read -d '' INTRO <<EOM
 ##############################################################################
 # install development-tools on linux (Thomas Schneider / 2016)
 # 
 # preconditions: debian 64bit system
-# PLEASE SET: PRJ, IP1, USER1, SHARE1 before starting
 ##############################################################################
+
+ _____     ______     __   __         ______     __  __     ______     __         __       
+/\  __-.  /\  ___\   /\ \ / /        /\  ___\   /\ \_\ \   /\  ___\   /\ \       /\ \      
+\ \ \/\ \ \ \  __\   \ \ \'/         \ \___  \  \ \  __ \  \ \  __\   \ \ \____  \ \ \____ 
+ \ \____-  \ \_____\  \ \__|          \/\_____\  \ \_\ \_\  \ \_____\  \ \_____\  \ \_____\
+  \/____/   \/_____/   \/_/            \/_____/   \/_/\/_/   \/_____/   \/_____/   \/_____/
+                                                                                           
+
 EOM
 
 clear
@@ -28,7 +36,9 @@ if [ "$DO_FORMAT" == "y" ]; then
 fi
 
 echo "Installing development like Java+Netbeans, Python+Anaconda and Sublime-Text"
-read -p "Virtualbox Version                            : " -i 5.1.6 VB_VERSION
+
+read -ep "Linux System Bit-width (32|64)                : " -i "64" BITS
+read -ep "Virtualbox Version                            : " -i 5.1.6 VB_VERSION
 read -p "Mount network-drive on IP                     : " IP1
 if [ "$IP1" != "" ]; then
     read -p "Mount network-drive on PATH                   : " SHARE1
@@ -55,30 +65,32 @@ sudo apt-get -y install vim ne dos2unix poppler-utils docx2txt catdoc colordiff 
 echo "install networking tools..."
 sudo apt-get -y install nmap git curl wget openssh-server openvpn links2 w3m tightvncserver
 
-echo "install virtualbox guest additions"
-#sudo apt-get -y install virtualbox-guest-utils virtualbox-guest-x11
-sudo apt-get -y install linux-headers-$(uname -r) build-essential dkms
-wget http://download.virtualbox.org/virtualbox/$VB_VERSION/VBoxGuestAdditions_$VB_VERSION.iso
-sudo mkdir /media/VBoxGuestAdditions
-sudo mount -o loop,ro VBoxGuestAdditions_$VB_VERSION.iso /media/VBoxGuestAdditions
-sudo sh /media/VBoxGuestAdditions/VBoxLinuxAdditions.run
-rm VBoxGuestAdditions_$VB_VERSION.iso
-sudo umount /media/VBoxGuestAdditions
-sudo rmdir /media/VBoxGuestAdditions
+if [ "$VB_VERSION" != "" ]; then
+	echo "install virtualbox guest additions"
+	#sudo apt-get -y install virtualbox-guest-utils virtualbox-guest-x11
+	sudo apt-get -y install linux-headers-$(uname -r) build-essential dkms
+	wget http://download.virtualbox.org/virtualbox/$VB_VERSION/VBoxGuestAdditions_$VB_VERSION.iso
+	sudo mkdir /media/VBoxGuestAdditions
+	sudo mount -o loop,ro VBoxGuestAdditions_$VB_VERSION.iso /media/VBoxGuestAdditions
+	sudo sh /media/VBoxGuestAdditions/VBoxLinuxAdditions.run
+	rm VBoxGuestAdditions_$VB_VERSION.iso
+	sudo umount /media/VBoxGuestAdditions
+	sudo rmdir /media/VBoxGuestAdditions
+fi
 
 if [ "$INST_NETBEANS" != "n" ]; then
     echo "install java+netbeans..."
     # wget http://download.oracle.com/otn-pub/java/jdk/8u112-b15/jdk-8u112-linux-i586.tar.gz
-    wget --no-check-certificate --no-cookies --header "Cookie: oraclelicense=accept-securebackup-cookie" http://download.oracle.com/otn-pub/java/jdk-nb/8u111-8.2/jdk-8u111-nb-8_2-linux-x64.sh
-    sudo bash jdk-8u111-nb-8_2-linux-x64.sh --silent &
-    jdk-8u111-nb-8_2-linux-x64.sh
+    wget --no-check-certificate --no-cookies --header "Cookie: oraclelicense=accept-securebackup-cookie" http://download.oracle.com/otn-pub/java/jdk-nb/8u111-8.2/jdk-8u111-nb-8_2-linux-x$BITS.sh
+    sudo bash jdk-8u111-nb-8_2-linux-x$BITS.sh --silent &
+    jdk-8u111-nb-8_2-linux-x$BITS.sh
     wget http://plugins.netbeans.org/download/plugin/3380
 fi
 
 echo "install sublime-text + plugins..."
-wget https://download.sublimetext.com/sublime-text_build-3126_amd64.deb
-sudo dpkg-deb -x sublime-text_build-3126_amd64.deb /
-rm sublime-text_build-3126_amd64.deb
+wget https://download.sublimetext.com/sublime-text_build-3126_amd$BITS.deb
+sudo dpkg-deb -x sublime-text_build-3126_amd$BITS.deb /
+rm sublime-text_build-3126_amd$BITS.deb
 # start sublime-text to create the directory structure
 subl
 wget  https://packagecontrol.io/Package%20Control.sublime-package
@@ -114,11 +126,11 @@ EOM
 
 if [ "$INST_ANACONDA" != "n" ]; then
     echo "install python-anaconda..."
-    wget https://repo.continuum.io/archive/Anaconda3-4.2.0-Linux-x86_64.sh
+    wget https://repo.continuum.io/archive/Anaconda3-4.2.0-Linux-x86_$BITS.sh
     #sudo rm -rf anaconda3
     # the bash script provides -b for non-interactive - but with unwanted defaults
-    echo -e "\n yes\n\nyes\n" | bash Anaconda3-4.2.0-Linux-x86_64.sh
-    rm Anaconda3-4.2.0-Linux-x86_64.sh
+    echo -e "\n yes\n\nyes\n" | bash Anaconda3-4.2.0-Linux-x86_$BITS.sh
+    rm Anaconda3-4.2.0-Linux-x86_$BITS.sh
 fi
 
 #sudo apt -y install python3 python3-pip python3-nose
@@ -137,8 +149,8 @@ chmod a+x fzf-install.sh
 
 echo "installing micro editor"
 MICRO_DIR=micro-1.2.0
-wget https://github.com/zyedidia/micro/releases/download/v1.2.0/$MICRO_DIR-linux64.tar.gz
-tar -xvf $MICRO_NAME-linux64.tar.gz
+wget https://github.com/zyedidia/micro/releases/download/v1.2.0/$MICRO_DIR-linux$BITS.tar.gz
+tar -xvf $MICRO_NAME-linux$BITS.tar.gz
 cp $MICRO_DIR/micro bin/
 
 # ----------------------------------------------------
@@ -147,16 +159,16 @@ cp $MICRO_DIR/micro bin/
 
 if [ "$INST_RESILIO_SYNC" == "y" ]; then
     echo "install/run bittorrent-sync. use localhost:8888 to configure it"
-    wget https://download-cdn.resilio.com/stable/linux-x64/resilio-sync_x64.tar.gz
-    tar -xf resilio-sync_x64.tar.gz
+    wget https://download-cdn.resilio.com/stable/linux-x$BITS/resilio-sync_x$BITS.tar.gz
+    tar -xf resilio-sync_x$BITS.tar.gz
     ./rslsync
 fi
 
 if [ "$IP1" != "" ]; then
     echo "domain"
-    wget http://download.beyondtrust.com/PBISO/8.0.0.2016/linux.deb.x64/pbis-open-8.0.0.2016.linux.x86_64.deb.sh
-    chmod +x pbis-open-8.0.0.2016.linux.x86_64.deb.sh
-    sudo ./pbis-open-8.0.0.2016.linux.x86_64.deb.sh
+    wget http://download.beyondtrust.com/PBISO/8.0.0.2016/linux.deb.x$BITS/pbis-open-8.0.0.2016.linux.x86_$BITS.deb.sh
+    chmod +x pbis-open-8.0.0.2016.linux.x86_$BITS.deb.sh
+    sudo ./pbis-open-8.0.0.2016.linux.x86_$BITS.deb.sh
     domainjoin-cli join $DOMAIN $DOMAIN_USER
 
     echo "connect network share drives"
@@ -170,9 +182,11 @@ if [ "$IP1" != "" ]; then
     bin/mount-$SHARE1
 fi
 
-echo "prepare ssh key to be copied to server machines"
-echo -e "\n\n\n" | ssh-keygen -t rsa
-cat ~/.ssh/id_rsa.pub | xclip -sel clip
+if [ ! -f ~/.ssh/id_rsa.pub ]; then 
+	echo "prepare ssh key to be copied to server machines"
+	echo -e "\n\n\n" | ssh-keygen -t rsa
+	cat ~/.ssh/id_rsa.pub | xclip -sel clip
+fi
 
 if [ "$REPO" != "" ]; then
     echo "get git projects"
@@ -184,8 +198,10 @@ if [ "$REPO" != "" ]; then
     subl
 fi
 
-echo "package the box..."
-sudo apt-get clean
-sudo dd if=/dev/zero of=/EMPTY bs=1M
-sudo rm -f /EMPTY
-#sudo shutdown -h now
+if [ "$VB_VERSION" != "" ]; then
+	echo "package the box..."
+	sudo apt-get clean
+	sudo dd if=/dev/zero of=/EMPTY bs=1M
+	sudo rm -f /EMPTY
+	#sudo shutdown -h now
+fi
